Point Estimate Mass Balance

Function conducting a two component, two node, point estimate mass balance from (Wills 2006) on a two node process. This function is provided for the purpose of comparing the performance of a Bayesian mass balance to a point estimate mass balance using any output of twonodeSim.

pointmassbal(y)

Arguments

y
A list of matrices of observed mass flow rates. Each matrix is a separate sample component. The rows of each matrix index the sampling location, and the columns index the sample set number. Necessary to format exactly the same as the output of twonodeSim()$simulation. Any arguments to twonodeSim can be used.

Value

Returns a list of vectors with mass flow rates yhat and grades ahat. Similar format to argument y. The index of a vector in the output is equivalent to the index of a row in y.

References

Wills BA (2006). Mineral processing technology : an introduction to the practical aspects of ore treatment and mineral recovery. Butterworth-Heinemann, Oxford Boston. ISBN 978-0-7506-4450-1.

Examples

y <- twonodeSim()$simulation

yhat <- pointmassbal(y)$yhat