Steady State Estimate

Allows for the estimation of process steady state of a single stream for a process using flow rate data.

ssEst(y, BTE = c(500, 20000, 1), stationary = FALSE)

Arguments

y
Vector of mass flow rate observations. Must be specified sequentially with y[1] as the initial observation.

BTE
Numeric vector giving c(Burn-in, Total-iterations, and Every) for MCMC approximation of target distributions. The function BMB produces a total number of samples of \((T - B)/E\). \(E\) specifies that only one of every \(E\) draws are saved. \(E > 1\) reduces autocorrelation between obtained samples at the expense of computation time.

stationary
Logical indicating if stationarity will be imposed when generating posterior draws. See Details.

Value

Returns a list of outputs

samples

List of vectors containing posterior draws of model parameters

stationary

Logical indicating the setting of the stationary argument provided to the ssEst function

y

Vector of observations initially passed to the ssEst function.

type

Character string giving details of the model fit. Primarily included for use with plot.BayesMassBal

Details

The model of the following form is fit to the data:

$$y_t = \mu + \alpha y_{t-1} + \epsilon$$

Where \(\epsilon \mathcal{N}(0,\sigma^2)\) and \(t\) indexes the time step.

A time series is stationary, and predictable, when \(|\alpha|< 1\). Enforcing stationarity, using stationary = TRUE, is achieved by specifying a truncated normal prior distribution for \(\alpha\) with a mean of zero and a variance of 10,000. Because \(\alpha\) and \(\mu\) are drawn together, a normal prior distribution with a mean of mean(y) and a variance of var(y)*1000.

When fitting a model where stationarity is not enforced, the Jeffreys prior of \(p(\mu,\alpha)\propto 1\) is used.

The Jeffreys prior of \(p(\sigma^2)\propto 1/\sigma^2\) is used for all inference of \(\sigma^2\)

A stationary time series will have an expected value of:

$$\frac{\mu}{1-\alpha}$$

Samples of this expectation are included in the output if stationary = TRUE or if none of the samples of \(alpha\) lie outside of (-1,1).

Examples

## Generating Data
y <- rep(NA, times = 21)

y[1] <- 0
mu <- 3
alpha <- 0.3
sig <- 2
for(i in 2:21){
 y[i] <- mu + alpha*y[i-1] + rnorm(1)*sig
}

## Generating draws of model parameters

ssEst(y, BTE = c(100,500,1))

#> $samples
#> $samples$mu
#>   [1]  2.60577057  3.43687722  4.02249158  4.74617953  3.19281696  3.15047737
#>   [7]  1.24627409  1.79010652  0.02738911  3.84241593  2.76851377  2.70129631
#>  [13]  4.17429138  3.11775244  2.65531193  3.66054290  1.64233695  2.66344837
#>  [19]  3.13231048  4.07777728  2.65565446  2.22366867  2.73964680  4.29655458
#>  [25]  2.98693163  3.90270502  3.02872440  3.17650495  5.55907976  1.64565774
#>  [31]  3.17845160  3.70375841  3.06544458  2.43653484  1.62893887  2.66886225
#>  [37]  3.68694210  4.74538975  3.93193989  3.03281157  4.07629287  3.44213428
#>  [43]  3.68467346  3.32573907  1.37868663  2.68236321  1.60294891  3.00434039
#>  [49]  3.42724570  3.22056583  4.14798074  1.53071930  3.19545085  3.05938625
#>  [55]  1.70691209  0.42606671  3.79758438  3.49881969  2.78493611  2.16707648
#>  [61]  3.00252845  3.55639164  0.43278285  4.65121960  4.31205969  2.54191745
#>  [67]  4.23452939  4.33313886  3.15448419  2.39774031  2.33510213  2.02673782
#>  [73]  2.69757423  2.43927060  3.05827826  2.89951310  3.44825806  2.77939913
#>  [79]  5.16259954  2.69212571  5.05208047  4.17456534  4.35823553  1.10762722
#>  [85]  2.43547592  4.21010220  2.55826433  4.52830868  5.91274467  4.08930680
#>  [91]  2.73754884  2.95354908  3.20207490  3.46914022  2.43297369  3.26478223
#>  [97]  2.65407776  3.07364005  4.13599341  3.15617372  2.40379057  2.45709905
#> [103]  0.94621648  0.85904892  4.51475052  3.22916836  3.90106753  4.73549835
#> [109]  2.97257055  4.20056487  4.18463510  5.44759016  2.84164467  1.32743658
#> [115]  3.79016173  3.72612940  1.91027203  2.53457575  2.83881641  3.42614897
#> [121]  3.37346327  3.74622916  1.77791985  3.60601548  3.36706324  4.22324355
#> [127]  2.86492787  4.49228549  4.73174977  3.67718171  3.35138981  3.01007912
#> [133]  3.79835973  2.61068367  3.06924242  1.86836913  4.07734637  4.94806109
#> [139]  3.45657853  2.94370417  4.10788455  2.94326632  3.52999421  4.19633743
#> [145]  1.49669919  2.85993003  3.12791642  3.67652153  4.91171740  2.43784771
#> [151]  3.83858568  3.86023669  1.71504481  1.75547253  2.71139729  2.60475880
#> [157]  2.11609676  1.02603457  2.31258209  3.78974587  2.64309579  4.00476051
#> [163]  2.42506196  2.53688596  3.60418223  4.23830315  3.52322330  2.72343116
#> [169]  0.33724119  3.44966736  1.96255079  2.51612565  2.79942277  3.63262208
#> [175]  2.53138437  4.18387459  3.40499702  3.12937796  1.93581359  4.52298563
#> [181]  3.66495943  4.68360633  2.57165070  3.07454370  3.50978775  2.83598397
#> [187]  3.95184466  1.98277257  2.23269903  4.72774256  5.31600472  2.27702232
#> [193]  3.96525479  2.12434032  1.99758733  3.23162130  2.61263341  5.13920044
#> [199]  3.28718459  2.40151165  3.20933480  4.58022801  1.73797619  4.25928305
#> [205]  1.39694248  2.16973267  4.40226246  4.21900662  3.10566524  1.60935088
#> [211]  1.72239476  3.20001406  3.25376572  3.89767471  6.15766791  4.02276604
#> [217]  3.53410590  2.51769203  3.59178992  3.51345930  3.91224982  3.82295328
#> [223]  2.79207655  2.21947096  4.72922934  2.53112136  3.28998571  1.86989385
#> [229]  4.28973097  2.30571628  1.52358916  4.37259384  3.97112878  2.31798906
#> [235]  3.96718123  1.64252854  2.04242120  2.94791646  3.77436141  3.47472011
#> [241]  3.80923273  4.19757876  4.40233822  4.16189361  2.06000866  3.49182724
#> [247]  3.36353809  3.09158070  3.94579970  3.24597313  2.86395444  2.92486038
#> [253]  3.82222897  3.31579424  4.86358024  2.47726360  3.51807139  0.71990879
#> [259]  4.30329833  3.82463887  2.08835061  5.32459701  1.89012653  0.37613912
#> [265]  1.76376716  1.90806507  3.11379058  3.03026367  1.43916875  3.19456030
#> [271]  3.36526510  6.57357536  3.06974197  2.87554357  3.58351522  0.97801666
#> [277]  3.87750383  2.42678749  2.80648687  3.59683297  3.60952003  2.82631090
#> [283]  3.00772201  3.05596327  2.78855826  4.34377794  2.58069985  2.98738475
#> [289]  2.47269252  3.81335338  2.88669354  2.70706510  2.04769864  3.48048614
#> [295]  3.97282987  3.22993536  3.42953011  3.37614813  3.36332888  5.88834556
#> [301]  1.38364260  3.16633011  2.70353983  2.68190329  2.56157903  2.69339384
#> [307]  3.95131125  0.79212421  2.05914637  1.41037195  5.22023571  0.19712961
#> [313]  3.27232849  3.12079786 -0.11715444  4.89269259  2.00045970  3.10885146
#> [319]  3.37267292  3.12168170  3.41909081  3.40572064  4.16093598  0.77070853
#> [325]  3.76688148  2.42943434  2.17599030  1.92635736  2.27879261  2.59861852
#> [331]  2.95505552  2.01820886  0.08509531  1.90520252  2.46633162  1.83865517
#> [337]  2.50895373  4.03406173  3.46210419  3.35399088  5.81655338  2.57793815
#> [343]  3.15950343  2.59298190  3.88958711  2.46410180  4.17857652  4.63892685
#> [349]  1.79475849  2.31912256  2.31904491  1.44311146  3.26790040  2.49343907
#> [355]  4.73229696  2.53993197  4.53510985  3.03067232  2.66097825  2.44087206
#> [361]  3.88509633  3.26495451  2.81582160  1.73073837  3.08257401  4.15031116
#> [367]  1.57569773  2.74701442  1.83624386  1.44365530  3.51078172  4.07934423
#> [373]  5.36294750  2.78194381  5.30541016  2.40190452  1.11292277  3.88523294
#> [379]  1.80987219  0.67622890  2.23781311  4.77060990  2.36521259  2.29638530
#> [385]  3.89259573  2.67068606  2.80017635  2.06607257  3.00363628  3.43612169
#> [391]  2.21411713  2.15695864  2.92853630  3.05469080  4.04127252  1.78065195
#> [397]  2.57511048  4.52917661  4.64592669  3.25758530
#> 
#> $samples$alpha
#>   [1]  0.2452770465  0.2836561330  0.1876573271  0.0661703970  0.2381315333
#>   [6]  0.1546873165  0.6135013408  0.5684120416  0.5827438561  0.2305432883
#>  [11]  0.2408894223  0.2842650203  0.2270130049  0.2164095841  0.2197781552
#>  [16] -0.1075950490  0.3805918909  0.2895732831  0.1927809559  0.0722445170
#>  [21]  0.1432897779  0.3902662009  0.2402930979  0.1121610381  0.2066310828
#>  [26] -0.0744152143  0.1557523256  0.2920868602  0.0607078600  0.5187648606
#>  [31]  0.2335735441  0.2349867034  0.3467142833  0.3167489994  0.4459448210
#>  [36]  0.3345275377  0.3733575273 -0.0245095399 -0.0187991689  0.3841729406
#>  [41] -0.0631442551  0.0006329392 -0.0575919580  0.2184302134  0.6085963589
#>  [46]  0.3724225234  0.5308748230  0.3098128062  0.0301523835  0.2196555634
#>  [51]  0.0015728642  0.3842840889  0.2599893059  0.2841595425  0.4583299542
#>  [56]  0.7339369129  0.0232725470  0.1306956315  0.2240787232  0.4590978900
#>  [61]  0.0898893442  0.0154353211  0.4112015595  0.0433664212 -0.1313656422
#>  [66]  0.2606074977  0.0428404032 -0.0078888752 -0.0175675893  0.2669095515
#>  [71]  0.3356151109  0.4214238620  0.4034335649  0.3550862405  0.3281287662
#>  [76]  0.2512351113  0.3086924698  0.4140461024 -0.0453409414  0.1929329491
#>  [81] -0.2199241695 -0.0024755304  0.0084342761  0.7416688059  0.3181866351
#>  [86] -0.0750606059  0.1950539008 -0.0751674542 -0.0287030027 -0.0493190008
#>  [91]  0.1658800081  0.1760332152  0.0034841368  0.1158116466  0.4222765904
#>  [96]  0.0814016183  0.4390186816  0.3705075162  0.0461395590  0.1638054386
#> [101]  0.4792777702  0.2100308822  0.5247958600  0.2857323552 -0.1077760168
#> [106]  0.2043593016  0.0505030977 -0.1471270392  0.3621989872  0.0858633419
#> [111] -0.0595277611 -0.1452122599  0.2593446548  0.4974662600  0.1464963462
#> [116] -0.0298346243  0.2973190174  0.1995904202  0.1643702539  0.2157115770
#> [121]  0.2074439591  0.0117779396  0.4374588596 -0.0114878642  0.1128512249
#> [126] -0.0765973212  0.3718966928 -0.1280508205 -0.0791452723  0.1844020079
#> [131]  0.4873794476  0.1983840934  0.0902243519  0.2811960455  0.1312913066
#> [136]  0.6362630781 -0.0530727992 -0.1945668714  0.2732672871  0.2633501122
#> [141]  0.1378411919  0.4034028549 -0.0224627035  0.1717090653  0.5784905123
#> [146]  0.2597399744  0.2124235674  0.0920566313 -0.2665397266  0.3072013681
#> [151]  0.1914317895  0.1226032440  0.4375864669  0.4920099873  0.3172174158
#> [156]  0.4879024410  0.3572645813  0.6994359115  0.2560029271 -0.0410961288
#> [161]  0.2783675190  0.1569873592  0.3115290784  0.5085578499  0.2357474288
#> [166] -0.0770727732  0.1447905890  0.4684073340  0.9429753373  0.0794492620
#> [171]  0.3645955354  0.4369914912  0.2146189273  0.2208910410  0.4347755301
#> [176]  0.0505130042  0.1930323400  0.2504532668  0.6007180084 -0.2343068094
#> [181]  0.0807221942 -0.3028719284  0.2097078046  0.1722102392  0.0989527170
#> [186]  0.2817158953  0.0782696601  0.2282445733  0.2065180315  0.2106082847
#> [191] -0.1160038059  0.3712910852  0.3056050964  0.4804630960  0.4020818783
#> [196]  0.1648563459  0.1423685094 -0.0930268198  0.1245784669  0.2314528916
#> [201]  0.2374231947 -0.1224058858  0.5940445306 -0.0194070900  0.8070600216
#> [206]  0.3770101573  0.0253403436  0.0872356601  0.3375186790  0.5329156367
#> [211]  0.5926615109  0.2734912475  0.2615285906  0.2150558498 -0.5121018340
#> [216]  0.0989271208  0.1184562087  0.3320651672  0.0500736600  0.2876737042
#> [221] -0.1257917178  0.1796030738  0.5915477896  0.2337687158 -0.2091160267
#> [226]  0.2784195252  0.1686869981  0.4845069934 -0.1028877051  0.5048535402
#> [231]  0.2797452879  0.0903183006  0.1936001196  0.1502738162 -0.0017996729
#> [236]  0.6433505655  0.6134302583  0.3251364046 -0.0258074254  0.0981790854
#> [241] -0.1755484411  0.0385879802 -0.1070513160  0.2372816700  0.4036756662
#> [246]  0.0611643353  0.2080288657  0.2806092651  0.2990237868  0.1654308956
#> [251]  0.3640297974  0.2809364196  0.0013197526  0.2468024989  0.0542675341
#> [256]  0.3387528817  0.2550245170  0.9694648712  0.0804340146 -0.0383838311
#> [261]  0.3348698857 -0.3464893942  0.3719393758  0.6243070996  0.4992941035
#> [266]  0.4119528751  0.2081763465  0.3052770873  0.6339538188  0.0028512894
#> [271] -0.0657186094 -0.2191396420 -0.0161628797  0.0598325216  0.1414536325
#> [276]  0.5561144377 -0.0458950252  0.4435947728  0.1643160238 -0.0920170856
#> [281]  0.0252002897  0.3098197036  0.0731087377  0.1182643071  0.0503470774
#> [286]  0.1326149068  0.6124693571  0.1028935538  0.1310643539  0.1152462198
#> [291]  0.3144017516  0.3521909365  0.4837597161  0.2464866163  0.0307144758
#> [296]  0.0630753962  0.1982315329  0.0189736981  0.2055850426 -0.2747409815
#> [301]  0.6104836999  0.2508453312  0.5218350759  0.1642870468  0.3314248424
#> [306]  0.3077976435  0.1660526496  0.7971407128  0.4570470599  0.9036762440
#> [311] -0.1114761833  0.8113616920  0.1121131941  0.1007881676  0.3981396015
#> [316] -0.1604847090  0.5029279873  0.2729888607  0.2076744925  0.0875787023
#> [321]  0.3189266076 -0.0712088435  0.0666208000  0.4746796570 -0.0606690036
#> [326]  0.3489893323  0.3931018665  0.3769538404  0.3212516096  0.2947382745
#> [331]  0.4249144243  0.1452560492  0.8030216261  0.5313027205  0.5335563313
#> [336]  0.3267134246  0.3942360879  0.0291169059  0.1512452540 -0.0262065820
#> [341] -0.1343521456  0.2183442000  0.3701696008  0.2830356016  0.0292612254
#> [346]  0.4657595077 -0.0515296235  0.1822260248  0.2284714628  0.2314041602
#> [351]  0.3990961404  0.2383563501  0.3045646913  0.2809705595 -0.3243614966
#> [356]  0.4489131645  0.0315205838  0.1792881820  0.2811650038  0.3740073305
#> [361]  0.0380553171  0.1701028621  0.2158942465  0.5108054918  0.1503026567
#> [366] -0.0045635922  0.5942361976  0.2033609549  0.5751172829  0.3703705587
#> [371]  0.0823942734 -0.1371578404 -0.0513051950  0.2817112258 -0.1684820294
#> [376]  0.1787431037  0.6301892583 -0.0711078327  0.3929336026  0.4856259121
#> [381]  0.3413410152 -0.0643485213  0.4376275390  0.2739907324 -0.1078549707
#> [386]  0.3364083035  0.3717935742  0.3959445535  0.1826178369  0.0382380224
#> [391]  0.3510127433  0.3914242971  0.0892423586  0.2232222394 -0.0344276371
#> [396]  0.4682247056  0.3638722305 -0.0349838909 -0.1228236290  0.1942671516
#> 
#> $samples$expectation
#>   [1]  3.4526187  4.7978037  4.9517177  5.0824899  4.1907719  3.7269965
#>   [7]  3.2245237  4.1477212  0.0656410  4.9936739  3.6470494  3.7741572
#>  [13]  5.4002091  3.9788037  3.4032781  3.3049470  2.6514618  3.7490825
#>  [19]  3.8803724  4.3953147  3.0998281  3.6469500  3.6061892  4.8393400
#>  [25]  3.7648710  3.6323993  3.5874833  4.4871394  5.9183714  3.4196541
#>  [31]  4.1471058  4.8414301  4.6923490  3.5660904  2.9400300  4.0104774
#>  [37]  5.8836454  4.6318649  3.8593866  4.9247780  3.8341860  3.4443143
#>  [43]  3.4840218  4.2552042  3.5224165  4.2741547  3.4168895  4.3529356
#>  [49]  3.5337981  4.1271081  4.1545152  2.4860805  4.3181144  4.2738381
#>  [55]  3.1512027  1.6013747  3.8880697  4.0248500  3.5891993  4.0064116
#>  [61]  3.2990806  3.6121463  0.7350272  4.8620702  3.8113759  3.4378459
#>  [67]  4.4240578  4.2992228  3.1000242  3.2707292  3.5146828  3.5029751
#>  [73]  4.5218337  3.7823206  4.5518815  3.8723946  4.9880233  4.7433751
#>  [79]  4.9386753  3.3356903  4.1413070  4.1642566  4.3953068  4.2876247
#>  [85]  3.5720566  3.9161534  3.1781809  4.2117241  5.7477665  3.8971055
#>  [91]  3.2819605  3.5845487  3.2132704  3.9235308  4.2113123  3.5540910
#>  [97]  4.7311339  4.8827272  4.3360572  3.7744490  4.6162626  3.1103736
#> [103]  1.9911789  1.2026989  4.0755084  4.0585762  4.1085627  4.1281377
#> [109]  4.6606551  4.5951170  3.9495285  4.7568388  3.8366626  2.6414875
#> [115]  4.4407094  3.6181823  2.7185481  3.1665985  3.3972180  4.3684809
#> [121]  4.2564350  3.7908779  3.1605152  3.5650606  3.7953761  3.9227699
#> [127]  4.5612367  3.9823432  4.3847199  4.5085713  6.5377594  3.7550142
#> [133]  4.1750510  3.6319829  3.5331089  5.1365947  3.8718561  4.1421382
#> [139]  4.7563272  3.9960695  4.7646495  4.9334234  3.4524430  5.0662602
#> [145]  3.5508078  3.8634128  3.9715719  4.0492851  3.8780603  3.5188402
#> [151]  4.7473863  4.3996478  3.0494373  3.4557225  3.9710991  5.0864503
#> [157]  3.2923295  3.4136965  3.1083215  3.6401498  3.6626619  4.7505344
#> [163]  3.5223884  5.1621253  4.7159570  3.9350202  4.1197200  5.1231541
#> [169]  5.9139533  3.7473951  3.0886638  4.4690721  3.5644133  4.6625341
#> [175]  4.4785470  4.4064580  4.2194963  4.1750271  4.8482367  3.6643933
#> [181]  3.9867811  3.5948325  3.2540505  3.7141601  3.8952315  3.9482761
#> [187]  4.2874195  2.5691722  2.8137993  5.9890957  4.7634288  3.6217433
#> [193]  5.7103743  4.0889113  3.3409045  3.8695394  3.0463357  4.7018063
#> [199]  3.7549734  3.1247423  4.2085398  4.0807234  4.2811991  4.1781964
#> [205]  7.2402956  3.4827738  4.5167176  4.6222299  4.6879288  3.4455251
#> [211]  4.2284115  4.4046463  4.4060822  4.9655440  4.0722574  4.4644181
#> [217]  4.0089964  3.7693678  3.7811247  4.9323734  3.4751098  4.6598825
#> [223]  6.8357484  2.8966071  3.9113114  3.5077465  3.9575776  3.6273894
#> [229]  3.8895446  4.6566349  2.1153477  4.8067295  4.9245156  2.7279247
#> [235]  3.9600544  4.6054427  5.2834482  4.3681664  3.6794054  3.8530046
#> [241]  3.2403877  4.3660560  3.9766343  5.4566587  3.4545105  3.7193168
#> [247]  4.2470463  4.2974986  5.6290066  3.8894001  4.5032840  4.0675963
#> [253]  3.8272800  4.4022905  5.1426597  3.7463507  4.7223989 23.5764123
#> [259]  4.6797059  3.6832612  3.1397625  3.9544292  3.0094651  1.0011877
#> [265]  3.5225612  3.2447486  3.9324294  4.3618306  3.9316590  3.2036950
#> [271]  3.1577426  5.3919790  3.0209153  3.0585440  4.1739332  2.2033081
#> [277]  3.7073547  4.3615469  3.3583112  3.2937515  3.7028325  4.0950327
#> [283]  3.2449567  3.4658496  2.9363973  5.0079002  6.6593440  3.3300226
#> [289]  2.8456567  4.3100730  4.2104739  4.1788009  3.9665611  4.6190104
#> [295]  4.0987199  3.4473802  4.2774570  3.4414451  4.2337180  4.6192487
#> [301]  3.5522072  4.2265372  5.6539903  3.2091202  3.8314003  3.8910498
#> [307]  4.7380824  3.9047964  3.7924951 14.6419949  4.6966690  1.0450137
#> [313]  3.6855244  3.4705925 -0.1946538  4.2160767  4.0244867  4.2762088
#> [319]  4.2566759  3.4213161  5.0201503  3.1793246  4.4579266  1.4671210
#> [325]  3.5514203  3.7317888  3.5854292  3.0918373  3.3573451  3.6846158
#> [331]  5.1384622  2.3611853  0.4320033  4.0648892  5.2875230  2.7308656
#> [337]  4.1418012  4.1550437  4.0790396  3.2683389  5.1276435  3.2980478
#> [343]  5.0164353  3.6166118  4.0068319  4.6123456  3.9738077  5.6726272
#> [349]  2.3262373  3.0173499  3.8592611  1.8947331  4.6990717  3.4677844
#> [355]  3.5732668  4.6089505  4.6827117  3.6927363  3.7017929  3.8992023
#> [361]  4.0387939  3.9341677  3.5911248  3.5379350  3.6278494  4.1314569
#> [367]  3.8832881  3.4482548  4.3217664  2.2928650  3.8260242  3.5873158
#> [373]  5.1012280  3.8730159  4.5404294  2.9246689  3.0094387  3.6273033
#> [379]  2.9813414  1.3146636  3.3975292  4.4821877  4.2057760  3.1630248
#> [385]  3.5136329  4.0245923  4.4574144  3.4203360  3.6747025  3.5727361
#> [391]  3.4116496  3.5442733  3.2154946  3.9325158  3.9067716  3.3485045
#> [397]  4.0481026  4.3760842  4.1377172  4.0430092
#> 
#> $samples$s2
#>   [1]  5.330303  8.210299  8.601819  4.509798  5.454977  4.817110  6.858911
#>   [8] 10.988010  7.810607  4.385613  4.459286  6.916047  4.623230  4.508299
#>  [15]  5.681296  5.790953  5.403213  5.068174  4.107328  5.515200  9.798591
#>  [22]  3.907267  4.452461  3.964421  4.537299  5.813705  6.136710  9.922215
#>  [29] 12.174879  8.027744  5.784083  4.834841  3.638129 11.864980  4.145799
#>  [36]  4.577015  7.563351  5.651192 11.702344  5.984564  6.069509  6.951761
#>  [43] 10.032822  3.660955  3.703630  3.692998  6.778536  3.811121  4.653215
#>  [50]  5.107905  7.621589  5.812434  7.402946  4.689024  7.657390  5.872593
#>  [57]  6.776589  3.440937  5.540760  5.827352  5.663350  6.734775  7.653631
#>  [64]  8.484884  5.610880  6.701740  5.774258  8.661630  7.942407  7.703966
#>  [71]  5.926185  6.746796  4.298033  5.486293  7.020324  3.002450 10.231162
#>  [78]  4.027476  6.077026  6.713054  4.670432  3.675419  9.902603  4.855491
#>  [85]  5.731317  8.512044  7.988229 10.287723  8.830912  4.434235  3.864095
#>  [92]  4.228814  5.538992  5.124946  3.030368  6.640135  4.639374  5.565647
#>  [99]  6.901705  5.415938  6.205148  3.758500 11.279502  8.032179  7.027452
#> [106] 12.724792  7.111278 12.370515  5.194279  4.698742  5.875971  7.393361
#> [113]  4.100947  4.566817  5.161568  8.549065  5.324463  7.623103  5.586822
#> [120]  3.171274  4.986025  6.579948  4.029412  5.815148  5.542808 12.447872
#> [127]  7.627434  9.181845  3.709600 12.902185  6.639268  5.024539  4.918743
#> [134]  4.568846  9.691686  6.645254  4.447378  6.284711  6.160422  8.104737
#> [141]  4.897854  9.009888  5.967073  4.619003  4.320172  6.903247  5.198625
#> [148] 10.653207 10.390047  4.567096  5.977496  5.121677  4.889990  4.107306
#> [155]  9.422001  4.978418 12.341658 10.465438  5.569377  8.440844  5.766096
#> [162]  6.634407  5.968097 10.745349  8.034491  4.730937  7.702068  9.161007
#> [169]  8.086508  5.135277  7.846217  9.595317  6.617068  8.886204  5.349736
#> [176]  5.478317  5.133667  4.290916  7.168829  4.520799  5.834876  5.688334
#> [183]  3.213567  6.166015  4.717863 10.380112  3.903468 11.099853  4.752334
#> [190] 10.420112  4.106329  4.966210  6.287455  7.311972  5.043279  4.812233
#> [197]  7.227985  6.542278  4.774255  5.819154  6.302811  3.886650  2.924707
#> [204]  6.418716  6.127057  8.064546  7.450967  5.002870  4.232107  7.127251
#> [211]  4.552178  6.491150  4.755670  4.413268  7.840319  4.152888  3.014215
#> [218]  7.503899  8.604753  6.802452  4.897800 12.435062  6.868331  5.990046
#> [225]  6.916302  3.086377 11.951932  3.884820  6.663221  4.459709 11.126699
#> [232]  3.956482  8.510612  8.845985  9.759225  6.543833  7.505085  5.164861
#> [239]  3.728689  6.700731  5.635334  6.381477  9.598359  3.776089  7.576506
#> [246]  8.327735  3.053672  5.155593  6.506831  4.069173  3.751195  5.846045
#> [253]  5.271399  6.467505  8.300092  5.226451  5.754419  7.913838  6.233753
#> [260]  6.055341  7.177902  7.082647 10.784897  6.603599  5.014253  7.191947
#> [267]  4.790420  5.479242  6.115971  9.093489 18.005038  8.413562  4.346376
#> [274]  4.739041  8.019588  7.405465  3.683871  5.709586  5.116423  5.032142
#> [281]  4.940896  4.513909  6.036226  6.860572  8.363234  4.405369  5.244870
#> [288] 11.597148  6.997358  4.915726  4.992231  4.477093  6.619255  6.774664
#> [295]  4.417564  4.525200  3.821435  4.239658  4.819649  7.249102  5.608093
#> [302]  5.222627  5.680741  6.727105  3.701721  8.216727  8.483067  6.394626
#> [309]  9.279376 11.541716  4.021943  3.626873  2.885690 11.434005 19.140470
#> [316]  5.207967  7.623156  5.471616  4.393482 12.347238  5.618540  6.935009
#> [323] 18.513203 10.435198  8.125138  3.203666  5.029106  5.484107  8.259718
#> [330]  5.156798  4.338821 13.363250  7.529748  7.589544 10.795956  9.198735
#> [337] 11.370632  3.837546 10.282770  9.557881  6.917238  3.432445  4.593446
#> [344]  5.035299  5.718836  5.990977  5.610989 10.632323 10.489851  5.447165
#> [351] 10.994405  9.401916 10.824937  6.402169 12.875044  5.893616  5.029238
#> [358]  5.908389  4.927589  7.457864  7.323452  3.942773  6.071983  4.927991
#> [365]  4.077689  6.389899  5.595258  5.025318  4.164692  8.192270  5.016514
#> [372]  7.265246 14.373324  8.147536  8.378923  7.642032  6.097911  5.357321
#> [379]  6.884128  5.707821  4.161236  7.267198  9.530108  6.484729  6.420817
#> [386]  4.612010  8.121583  4.322933  3.271782  5.451266  5.472702  5.640769
#> [393]  5.963794  4.807909  5.017572  4.918370  7.357823  8.511956  4.764671
#> [400]  3.620654
#> 
#> 
#> $stationary
#> [1] FALSE
#> 
#> $y
#>  [1]  0.0000000  6.2243990  3.1876450  5.2008741  1.3742456  4.1163238
#>  [7]  6.3478864  6.7970360  8.3429274  6.6173682  4.8213194  4.8152381
#> [13]  0.7034014  3.4841804  5.7159678  1.4744640  2.6110160  4.5341237
#> [19] -0.5539306  1.7281433  2.2520146
#> 
#> $type
#> [1] "time-series"
#> 
#> attr(,"class")
#> [1] "BayesMassBal"