Summary of BayesMassBal Object

Prints a summary table containing mean values and 95

# S3 method for BayesMassBal
summary(object, export = NA, ...)

Arguments

object
A BayesMassBal object returned from the BMB function

export
Optional character string specifying location to save a *.csv file containing summary data. Only data related to mass flow rates is printed.

...
Additional arguments affecting the summary produced. Not used for a BayesMassBal object.

Value

A summary table printed to the console, and optionally a saved *.csv file saved within the path as specified.

Details

Current implementation only returns statistics for balanced mass flow rates, taken from x$ybal, and not statistics on \(\beta\) or variance parameters of \(\sigma^2\) and \(\Sigma\).

The header entry of the table 95% LB should be interpreted as the lower bound of the 95