Bayesian Mass Balance

Allows the user to specify the covariance structure for a Bayesian mass balance, simulates draws from reconciled masses and relevant covariance matrix, and approximates the log-marginal likelihood.

BMB(
  X,
  y,
  cov.structure = c("indep", "component", "location"),
  priors = NA,
  BTE = c(500, 20000, 1),
  lml = FALSE,
  ybal = TRUE,
  verb = 1
)

Arguments

X
A matrix that maps constrained masses to observed masses. Can be built from the function constrainProcess, see documentation for details.

y
A list of matrices of observed mass flow rates. Each matrix is a separate sample component. The rows of each matrix index the sampling location, and the columns index the sample set number. Can be specified using the importObservations function.

cov.structure
Character string. "indep" allows for independent error. "component" indicates error for a single sample component, between locations, is correlated. "location" indicates error all observations at a single location, between sample components, is correlated. Not specifying cov.structure defaults to the "indep" structure.

priors
Optional list of user specified hyperparameters for conjugate priors. When not specified, the function uses a set of default conjugate priors. To see the required list structure run BMB with BTE = c(1,2,1) and inspect the output. See Details for more information.

BTE
Numeric vector giving c(Burn-in, Total-iterations, and Every) for MCMC approximation of target distributions. The function BMB produces a total number of samples of \((T - B)/E\). \(E\) specifies that only one of every \(E\) draws are saved. \(E > 1\) reduces autocorrelation between obtained samples at the expense of computation time.

lml
Logical indicating if the log-marginal likelihood should be approximated. Default is FALSE, which reduces computation time. Log-marginal likelihood is approximated using methods in (Chib 1995) .

ybal
Logical indicating if the mass balanced samples for each \(y\) should be returned. Default is TRUE. Setting ybal=FALSE results in a savings in RAM and computation time.

verb
Numeric indicating verbosity of progress printed to R-console. The default of 1 prints messages and a progress bar to the console during all iterative methods. verb = 0 indicates no messages are printed.

Value

Returns a list of outputs

beta

List of matrices of samples from the distribution of reconciled data. Each matrix in the list is a separate sample component. Each column of a matrix in beta is a draw from the target distribution.

Sig

List of matrices containing draws from the distribution of each covariance matrix. If S.t is the \(t^{th}\) draw from the distribution of covariance matrix S and:

cov.structure = "component" or "location", the \(t^{th}\) column of a matrix in Sig is equal to S.t[upper.tri(W, diag = TRUE)].

priors

List of prior hyperparameters used in generating conditional posterior distributions and approximating log-marginal likelihood. The structure of the input argument priors is required to be the same as the structure of this returned list slice. See Details.

cov.structure

Character string containing the covariance structure used.

y.cov

List of character matrices indicating details for the structure of each covariance matrix. Only returned when cov.structure = "location"

lml

Numeric of the log-marginal likelihood approximation. Returns NA when lml = FALSE

ybal

List of samples from the distribution of reconciled mass flow rates, in the same format as the function argument y. Produced with argument ybal = TRUE. Equivalent to lapply(BMB(...)$beta,function(X,x){x %*% X} , x = X) . Viewing this output is more intuitive than viewing samples of beta, at the expense of RAM and some computation time.

X

The function argument X is passed to the output so that it can be used with other BayesMassBal functions.

Details

See vignette("Two_Node_Process", package = "BayesMassBal") for further details on how to use function outputs.

When the priors argument is left unspecified, a set of default conjugate priors are used, which are chosen to provide little information to the posterior distribution when adequate data is supplied. In the current version of the BayesMassBal package, only the conjugate priors stated below can be used, but hyperparameter values can be specified by the user.

The prior distribution on beta is a normal distribution truncated at 0. The mean of this distribution, before truncation, is 0 for each element of beta. Each beta is independent with a variance of 10,000,000 before truncation.

When cov.structure = "indep" all observations in a sample set are independent. A Gamma prior distribution, with \(a = 1\) and \(b = 100,000,000\), is placed on the precision of the mass flow rate for each sample component at each sample location. Parameterization used for the Gamma distribution is:

$$f(\phi) = \frac{b^a}{\Gamma(a)}\phi^{a-1}e^{-b\phi}$$

When cov.structure = "component" or "location", the prior distribution on \(\Sigma_i\) is inverse Wishart\((\nu_0, \nu_0 \times S_0)\). The degrees of freedom parameter, \(\nu_0\), is equal to the dimension of \(\Sigma_i\). The scale matrix parameter is equal to a matrix, \(S_0\), with the sample variance of the observation being correlated on the diagonal, multiplied by \(\nu_0\).

The user is able to specify the prior hyperparameters of the mean and variance of beta, \(a\) and \(b\) for each \(\phi\), and the degrees of freedom and scale matrix for each \(\Sigma_i\).

References

Chib S (1995). “Marginal likelihood from the Gibbs output.” Journal of the american statistical association, 90(432), 1313--1321. Casella G, George EI (1992). “Explaining the Gibbs sampler.” The American Statistician, 46(3), 167--174.

Examples

y <- importObservations(file = system.file("extdata", "twonode_example.csv",
                                    package = "BayesMassBal"),
                   header = TRUE, csv.params = list(sep = ";"))

C <- matrix(c(1,-1,0,-1,0,0,1,-1,0,-1), byrow = TRUE, ncol = 5, nrow = 2)
X <- constrainProcess(C = C)

BMB_example <- BMB(X = X, y = y, cov.structure = "indep",
                   BTE = c(10,300,1), lml = FALSE, verb=0)