Bayesian Mass Balance

One function that allows the user to specify covariance structure for a Bayesian mass balance, simulates draws from reconciled masses and relevant covariance matrix, and approximates the log(marginal likelihood).

BMB(
  X,
  y,
  cov.structure = c("indep", "component", "location"),
  priors = NA,
  BTE = c(500, 20000, 1),
  lml = FALSE,
  ybal = TRUE,
  verb = 1
)

Arguments

X
A matrix that maps constrained masses to observed masses. Can be built from the function constrainProcess, see documentation for details.

y
A list of matrices of observed mass flow rates. Each matrix is a separate sample component. The rows of each matrix index sampling location, and the columns index sample set number.

cov.structure
Character string. "indep" allows for independent error. "component" indicates error for a sample component is correlated, and error between components is independent. "location" indicates correlated error between all sample components at a given location, with independent error between location. Not specifying cov.structure defaults to the "indep" structure.

priors
Optional list of user specified hyperparameters for conjugate priors. To see the required list structure run BMB with BTE = c(1,2,1) and inspect the output. When not specified, the function defaults to the priors: \(p(\beta) = \) Truncated Normal \((0,10000000)\), for independent error \(p(\phi_{i,j}) = \) Gamma \((1,50000000)\), for location and component error \(p(\Sigma_i)\) Inverse Wishart\((\nu_0,\nu_0 \times S_0)\) where \(S_0\) is a diagonal matrix containing the empirical variance of the observations for each location and component, and \(\nu_0\) is equal to the dimension of \(\Sigma_i\).

BTE
Numeric vector giving c(Burn-in, Total-iterations, and Every) for MCMC approximation of target distributions. The function BMB produces a total number of samples of \((T - B)/E\). \(E\) specifies only one of Every \(E\) draws are saved, reduces autocorrelation between consecutive samples at the expense of computation time.

lml
Logical indicating if the log marginal likelihood should be approximated. Default is FALSE, which reduces computation time. Log-marginal likelihood is approximated using methods in (Chib 1995) .

ybal
Logical indicating if the mass balanced samples for each \(y\) should be returned. Default is TRUE, setting ybal=FALSE results in a savings in RAM and computation time.

verb
Numeric indicating verbosity of progress printed to R-console. 1 prints messages and a progress bar for all iterative methods, 0 indicates no verbosity.

Value

Returns a list of outputs

beta

List of matrices of samples from the distribution of reconciled data. Each matrix in the list is a separate sample component.

Sig

List of matrices containing draws from the distribution of each covariance matrix. If cov.structure = "indep" the elements of each matrix are the diagonal elements of the covariance between sample locations for a given component. For cov.structure = "component" or "location", if S.t is a draw from the distribution of covariance matrix S, the \(t^{th}\) column in a listed matrix is equal to S.t[upper.tri(W, diag = TRUE)]

priors

List of prior hyperparameters used in generating conditional posterior distributions and approximating log marginal likelihood. The structure of the input variable priors is required to be the same as the structure of the returned prior hyperparameters. See Details.

cov.structure

Character string containing the covariance structure used.

y.cov

List of character matrices indicating details for the structure of each covariance matrix. Only returned when cov.structure = "location"

lml

Numeric of the log marginal likelihood approximation. Returns NA when lml = FALSE

ybal

List of samples from the distribution of reconciled mass flow rates, in the same format as the function argument y. Produced with argument ybal = TRUE. Equivalent to lapply(BMB$beta,function(X,x){x %*% X} , x = X) . Viewing this output is more intuitive than viewing samples of beta, at the expense of RAM and some computation time.

X

The function argument X is passed to the output so that it can be used with other functions.

Details

See vignette("Two_Node_Process", package = "BayesMassBal") for further details on how to use function outputs.

References

Chib S (1995). “Marginal likelihood from the Gibbs output.” Journal of the american statistical association, 90(432), 1313--1321. Casella G, George EI (1992). “Explaining the Gibbs sampler.” The American Statistician, 46(3), 167--174.

Examples

## Importing Data
y <- importObservations(file = system.file("extdata", "twonode_example.csv",
                                    package = "BayesMassBal"),
                   header = TRUE, csv.params = list(sep = ";"))

C <- matrix(c(1,-1,0,-1,0,0,1,-1,0,-1), byrow = TRUE, ncol = 5, nrow = 2)
X <- constrainProcess(C = C)

BMB_example <- BMB(X = X, y = y, cov.structure = "indep",
                   BTE = c(100,3000,1), lml = FALSE, verb=0)